Print out article numbers as unsigned long with appropriate casts.   Thanks to Piotr Meyer for pointing out places where INN was assuming signed article numbers.

+++ nnrpd/article.c
@@ -647,7 +647,8 @@ void CMDfetch(int ac, char *av[])
 	    return;
 	}
 	tart=art;
-	Reply("%d %ld %s %s\r\n", what->ReplyCode, art, av[1], what->Item);
+	Reply("%d %lu %s %s\r\n", what->ReplyCode, (unsigned long) art,
+              av[1], what->Item);
 	if (what->Type != STstat) {
 	    ARTsendmmap(what->Type);
 	}
